git-svn-id: https://support.ginsys.be/svn/scripts@67 a33c6f8c-bf51-0410-aefe-ad234f6e076f

+++ virtest/upload.sh
@@ -1,4 +1,5 @@
 #!/bin/sh
+[ $# = 0 ] && { echo need target host; exit; }
 host=$1
 shift
 rsync -avL --delete $(dirname $0) root@$host:~/virtest/ --exclude=.swp $*
